Return all gadgets details to every user
Instead of returning only the gadgets the user has added
to his workspaces, now it returns all. It fixes a problem
when an administrator adds a new gadget to a shared workspace.
When that happens the workspace load crashes to every other
user which the workspace has been shared with, since
they won't receive the gadget details.